支持缩放(但是有些画面还是原本的画面, 而且使用线程安全的LRU似乎有bug)

+++ qt-client/include/mu/PageRenderer.h
@@ -84,7 +84,7 @@ protected:
 private:
     Document* _doc;
     int _pageNum;
-    int _zoom;
+    float _zoom;
     int _dpi;
 };
